Time: 84 ms (92.56%), Space: 57.5 MB (29.49%) - LeetHub

+++ 1814-count-nice-pairs-in-an-array/1814-count-nice-pairs-in-an-array.cpp
@@ -17,7 +17,7 @@ public:
                 ans-=mod; 
                 
         }
-        return (int)ans; 
+        return ans; 
     }
     int rev(int a)
     {
